test: migrate repository.rs unit tests to integration tests
Migrated all 31 tests from src/git/repository.rs to a new integration
test file tests/integration/repository_unit.rs. This includes:

- 19 pure unit tests that test parsing functions and profile rewriting
  without needing a repository at all
- 1 test (test_list_commit_files_with_utf8_filename) that was using
  TmpRepo, now rewritten to use TestRepo with CLI-based git operations
- 11 tests using raw git commands for repository operations

Made 6 previously private functions public to enable integration tests:
- args_with_disabled_hooks_if_needed
- args_with_internal_git_profile
- resolve_command_base_dir
- worktree_storage_ai_dir
- parse_git_version
- parse_diff_added_lines_with_insertions

All tests pass without modification. The one TmpRepo-based test was
successfully converted to use TestRepo with standard git CLI operations.
